+++ airflow/contrib/sensors/sagemaker_training_sensor.py
@@ -20,7 +20,7 @@
 import time
 
 from airflow.contrib.hooks.sagemaker_hook import LogState, SageMakerHook
-from airflow.contrib.sensors.sagemaker_base_sensor import SageMakerBaseSensor
+from airflow.providers.amazon.aws.sensors.sagemaker_base import SageMakerBaseSensor
 from airflow.utils.decorators import apply_defaults
